requirements.txt
aioredis==1.2.0
amqp==2.3.2
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
cffi==1.11.5
channels==2.1.5
channels-redis==2.3.1
constantly==15.1.0
daphne==2.2.3
Django==2.0.8
django-celery-beat==1.1.1
django-ckeditor==5.4.0
django-debug-toolbar==1.9.1
django-easy-timezones==0.8.0
django-js-asset==1.1.0
django-jsonfield==1.0.1
django-notify-x==0.1.9
djangorestframework==3.9.0
hiredis==0.3.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.16
kombu==4.2.1
lupa==1.4
msgpack==0.5.6
packaging==17.1
pbr==4.0.3
Pillow==5.1.0
pip-review==1.0
pycparser==2.18
pygeoip==0.3.2
PyHamcrest==1.9.0
pyparsing==2.2.0
pytz==2016.6
six==1.11.0
sqlparse==0.2.4
Twisted==18.9.0
txaio==18.8.1
stevedore==1.28.0
vine==1.1.4
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
zope.interface==4.6.0
